whole proj for integration/ust_employee_management -DRAFT/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.3.0
black==25.12.0
flake8==7.1.1
mypy==1.13.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
